test: 테스트 실행 환경 분리
