--- src/main/java/woowacourse/shoppingcart/dto/CustomerResponse.java
@@ -1,46 +1,0 @@
-package woowacourse.shoppingcart.dto;
-
-import woowacourse.shoppingcart.domain.customer.Customer;
-
-public class CustomerResponse {
-
-    private final String name;
-    private final String email;
-    private final String address;
-    private final String phoneNumber;
-
-    private CustomerResponse() {
-        this(null, null, null, null);
-    }
-
-    public CustomerResponse(String name, String email, String address, String phoneNumber) {
-        this.name = name;
-        this.email = email;
-        this.address = address;
-        this.phoneNumber = phoneNumber;
-    }
-
-    public static CustomerResponse from(Customer customer) {
-        return new CustomerResponse(
-            customer.getName(),
-            customer.getEmail(),
-            customer.getAddress(),
-            customer.getPhoneNumber());
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumber;
-    }
-}
