+++ src/test/java/woowacourse/shoppingcart/acceptance/CartAcceptanceTest.java
@@ -18,7 +18,8 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import woowacourse.shoppingcart.acceptance.fixture.CustomerAcceptanceFixture;
-import woowacourse.shoppingcart.domain.Cart;
+import woowacourse.shoppingcart.application.dto.CartItemResponse;
+import woowacourse.shoppingcart.application.dto.CartResponse;
 
 @DisplayName("장바구니 관련 기능")
 public class CartAcceptanceTest extends AcceptanceTest {
@@ -115,8 +116,10 @@ public class CartAcceptanceTest extends AcceptanceTest {
     }
 
     public static void 장바구니_아이템_목록_포함됨(ExtractableResponse<Response> response, Long... productIds) {
-        List<Long> resultProductIds = response.jsonPath().getList(".", Cart.class).stream()
-            .map(Cart::getProductId)
+        List<Long> resultProductIds = response.jsonPath().getObject(".", CartResponse.class)
+            .getItemResponses()
+            .stream()
+            .map(CartItemResponse::getProductId)
             .collect(Collectors.toList());
         assertThat(resultProductIds).contains(productIds);
     }
