+++ src/test/java/woowacourse/support/test/ExtendedJdbcTest.java
@@ -1,0 +1,20 @@
+package woowacourse.support.test;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import org.springframework.test.context.TestConstructor;
+import org.springframework.test.context.jdbc.Sql;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@JdbcTest
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+@Sql(scripts = {"classpath:schema-test.sql", "classpath:data-test.sql"})
+@TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
+public @interface ExtendedJdbcTest {
+}
