+++ src/test/resources/application.properties
@@ -1,3 +1,4 @@
+spring.sql.init.platform=test
 spring.sql.init.mode=embedded
 spring.datasource.url=jdbc:h2:~/test;MODE=MYSQL;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
 spring.datasource.username=sa
