+++ src/test/resources/data-test.sql
@@ -1,0 +1,26 @@
+insert into customer (name, password, email, address, phone_number)
+values ('puterism', 'abcd1234!', 'woowahan1@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('tanney-102', 'abcd1234!', 'woowahan2@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('jho2301', 'abcd1234!', 'woowahan3@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('365kim', 'abcd1234!', 'woowahan4@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('dudtjr913', 'abcd1234!', 'woowahan5@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('jum00', 'abcd1234!', 'woowahan6@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('hyuuunjukim', 'abcd1234!', 'woowahan7@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('zereight', 'abcd1234!', 'woowahan8@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('devhyun637', 'abcd1234!', 'woowahan9@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('swon3210', 'abcd1234!', 'woowahan10@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('bigsaigon333', 'abcd1234!', 'woowahan11@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('yungo1846', 'abcd1234!', 'woowahan12@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('zigsong', 'abcd1234!', 'woowahan13@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('iborymagic', 'abcd1234!', 'woowahan14@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('0307kwon', 'abcd1234!', 'woowahan15@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('gwangyeol-im', 'abcd1234!', 'woowahan16@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('shinsehantan', 'abcd1234!', 'woowahan17@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('ddongule', 'abcd1234!', 'woowahan18@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('seojihwan', 'abcd1234!', 'woowahan19@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('0imbean0', 'abcd1234!', 'woowahan20@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('sunyoungkwon', 'abcd1234!', 'woowahan21@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('hchayan', 'abcd1234!', 'woowahan22@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('2sooy', 'abcd1234!', 'woowahan23@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('yujo11', 'abcd1234!', 'woowahan24@techcourse.com', 'seongdam', '010-1234-5678'),
+      ('sunhpark42', 'abcd1234!', 'woowahan25@techcourse.com', 'seongdam', '010-1234-5678');
