+++ src/test/resources/schema-test.sql
@@ -11,13 +11,14 @@ drop table if exists customer;
 create table customer
 (
     id           bigint       not null auto_increment,
-    name     varchar(20) not null unique,
+    name         varchar(20)  not null unique,
     password     varchar(60)  not null,
     email        varchar(255) not null unique,
     address      varchar(255) not null,
     phone_number varchar(13)  not null,
     primary key (id)
-) engine=InnoDB default charset=utf8mb4;
+) engine = InnoDB
+  default charset = utf8mb4;
 
 create table product
 (
@@ -26,34 +27,28 @@ create table product
     price     integer      not null,
     image_url varchar(255),
     primary key (id)
-) engine=InnoDB default charset=utf8mb4;
+) engine = InnoDB
+  default charset = utf8mb4;
 
 create table cart_item
 (
     id          bigint not null auto_increment,
     customer_id bigint not null,
     product_id  bigint not null,
-    primary key (id)
-) engine=InnoDB default charset=utf8mb4;
-
-alter table cart_item
-    add constraint fk_cart_item_to_customer
-        foreign key (customer_id) references customer (id);
-
-alter table cart_item
-    add constraint fk_cart_item_to_product
-        foreign key (product_id) references product (id);
+    primary key (id),
+    foreign key (customer_id) references customer (id),
+    foreign key (product_id) references product (id)
+) engine = InnoDB
+  default charset = utf8mb4;
 
 create table orders
 (
     id          bigint not null auto_increment,
     customer_id bigint not null,
-    primary key (id)
-) engine=InnoDB default charset=utf8mb4;
-
-alter table orders
-    add constraint fk_orders_to_customer
-        foreign key (customer_id) references customer (id);
+    primary key (id),
+    foreign key (customer_id) references customer (id)
+) engine = InnoDB
+  default charset = utf8mb4;
 
 create table orders_detail
 (
@@ -61,13 +56,8 @@ create table orders_detail
     orders_id  bigint  not null,
     product_id bigint  not null,
     quantity   integer not null,
-    primary key (id)
-) engine=InnoDB default charset=utf8mb4;
-
-alter table orders_detail
-    add constraint fk_orders_detail_to_orders
-        foreign key (orders_id) references orders (id);
-
-alter table orders_detail
-    add constraint fk_orders_detail_to_product
-        foreign key (product_id) references product (id);
+    primary key (id),
+    foreign key (orders_id) references orders (id),
+    foreign key (product_id) references product (id)
+) engine = InnoDB
+  default charset = utf8mb4;
